easy-config/requirements.txt
loguru
tabulate
pyyaml
pydantic
msgpack